First submission of the incomplete project
Definitely looking for contributors on this one! See readme! It's already primed in-project to be ported by someone with experience! I'll keep working on it of course!

GitHub didn't allow the DATA_1.dat DATA_2.dat DATA_3.dat and DATA_4.dat files to be added because they were too large. I can send you them, or you can get them from a .XAP of the game, which would be a good reference file to have anyways.